+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+* Accept negation of grouped expression (example: `-(2 * 3)`)
+
 
 ## [2.0.3] - 2024-01-18